requirements.txt
fastapi >= 0.115.0
uvicorn >= 0.30.0
sqlalchemy >= 2.0.0
pydantic >= 2.0.0
psycopg2-binary >=  2.9
pydantic-settings>=2.0.0
python-decouple>=3.8

scikit-learn >= 1.5.0
joblib >= 1.4.0
sentence-transformers >= 3.0.0
torch >= 2.0.0
numpy >= 1.26.0
qdrant-client >= 1.10.0

streamlit >= 1.38.0
requests >= 2.32.0
pandas >= 2.2.0
nltk >= 3.8.0